ImageDownloaderHelper: Revise image detection
Not only use the first <img> tag, but use all of them in the event
that the first one wasn't a desired image. We will also check if
the image URL from the tag ends with a supported extension and that
it isn't a sprite sheet. Finally, we will also use a "legitimate"
user agent so that websites will be kind and not outright reject
requests that have a blank user agent.